Implement parseImport in lila.core.game
