+++ modules/study/src/test/Helpers.scala
@@ -12,10 +12,9 @@ object Helpers:
   import lila.tree.NewTree.*
 
   /* Remove these when tests are independant from lila.game.importer */
-  val parseImportStub: lila.core.game.ParseImport = (_, _) => Left(chess.ErrorStr("lila.game not available"))
-  val statusTextStub: lila.core.game.StatusText   = (_, _, _) => ""
-  val importerStub                                = StudyPgnImport(parseImportStub, statusTextStub)
-  val newImporterStub                             = StudyPgnImportNew(parseImportStub, statusTextStub)
+  val statusTextStub: lila.core.game.StatusText = (_, _, _) => ""
+  val importerStub                              = StudyPgnImport(statusTextStub)
+  val newImporterStub                           = StudyPgnImportNew(statusTextStub)
 
   def rootToPgn(root: Root): PgnStr = PgnDump
     .rootToPgn(root, Tags.empty)(using PgnDump.WithFlags(true, true, true, true, false))
